LPD-67215 Move to item-selector-web module
