+++ modules/apps/fragment/fragment-impl/src/main/resources/META-INF/resources/js/api/api.ts
@@ -8,9 +8,9 @@ export {LocalizationSelect} from './LocalizationSelect';
 export {focusInput} from './focusInput';
 export {getTranslationInput} from './getTranslationInput';
 export {handleInputLengthError, showInputError} from './handleInputError';
-export {openCMSItemSelectorModal} from './openCMSItemSelectorModal';
 export {registerLocalizedInput} from './registerLocalizedInput';
 export {registerUnlocalizedInput} from './registerUnlocalizedInput';
 export {updateDLVideo} from './updateDLVideo';
+export {openCMSItemSelectorModal} from '@liferay/frontend-js-item-selector-web';
 
 export {localStorage} from 'frontend-js-web';
