+++ modules/apps/frontend-js/frontend-js-item-selector-web/src/main/resources/META-INF/resources/index.ts
@@ -8,6 +8,5 @@ export type {IItemSelectorProps} from './item_selector/ItemSelector';
 export {default as ItemSelectorModal} from './item_selector/ItemSelectorModal';
 export type {FilesUploaderComponent} from './item_selector/ItemSelectorModal';
 export type {IItemSelectorModalProps} from './item_selector/ItemSelectorModal';
+export {default as openCMSItemSelectorModal} from './item_selector/openCMSItemSelectorModal';
 export {default as openItemSelectorModal} from './item_selector/openItemSelectorModal';
-
-export {default as CMSFileUploaderComponent} from './item_selector_file_uploader/CMSFileUploaderComponent';
